Revert "abstracted validated input out of guess input. There is still a bug but it is covered by a test."
This reverts commit b17e078a4402d44c4d0d55acff31982ffe8b4107.
