--- src/components/input/constants.js
@@ -1,6 +1,0 @@
-export const WARNINGS = {
-  TOO_SHORT: 'Too short',
-  TOO_LONG: 'Too long',
-  NO_DUPLICATES: 'No duplicates',
-  LETTERS_SIZE: 'Must be between 0-4',
-}
