+++ src/components/input/user-input.js
@@ -1,16 +1,18 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 
-import { ValidatedInput } from '../validated-input/validated-input'
-import { WARNINGS } from '../constants'
+import { hasDuplicates, checkLength } from './validation'
 
-const GuessInput = ({ addGuess, autocomplete, myWord }) => {
+const UserInput = ({ addGuess, autocomplete, myWord }) => {
   const [word, setWord] = useState('')
-  const [warning, setWarning] = useState('')
   const [letters, setLetters] = useState(0)
+  const [warning, setWarning] = useState('')
 
-  const setGuess = newWord => {
+  const updateWord = event => {
+    const newWord = event.target.value
+
     setWord(newWord)
+    setWarning(getWarning(newWord, letters))
 
     if (autocomplete && myWord) {
       setLetters(lettersInCommon(newWord, myWord))
@@ -19,15 +21,9 @@ const GuessInput = ({ addGuess, autocomplete, myWord }) => {
 
   const updateLetters = event => {
     const newLetters = event.target.value
-    setLetters(newLetters)
-    setWarning(getWarning(newLetters))
-  }
 
-  function getWarning(letters) {
-    if (letters < 0 || letters > 4) {
-      return WARNINGS.LETTERS_SIZE
-    }
-    return warning
+    setLetters(newLetters)
+    setWarning(getWarning(word, newLetters))
   }
 
   function lettersInCommon(word1, word2) {
@@ -37,6 +33,26 @@ const GuessInput = ({ addGuess, autocomplete, myWord }) => {
       }
       return total
     }, 0)
+  }
+
+  function getWarning(word, letters) {
+    if (!word) {
+      return ''
+    }
+    if (checkLength(word) === -1) {
+      return WARNINGS.TOO_SHORT
+    }
+    if (checkLength(word) === 1) {
+      return WARNINGS.TOO_LONG
+    }
+    if (hasDuplicates(word)) {
+      return WARNINGS.NO_DUPLICATES
+    }
+    if (letters < 0 || letters > 4) {
+      return WARNINGS.LETTERS_SIZE
+    }
+
+    return ''
   }
 
   function onSubmit(event) {
@@ -52,7 +68,13 @@ const GuessInput = ({ addGuess, autocomplete, myWord }) => {
   return (
     <div>
       <form onSubmit={onSubmit}>
-        <ValidatedInput setWarning={setWarning} setGuess={setGuess} />
+        <input
+          placeholder='Enter word...'
+          type='text'
+          onChange={updateWord}
+          value={word}
+          aria-label='Enter guess here'
+        />
         <input
           data-testid='letters_input'
           type='number'
@@ -74,10 +96,17 @@ const GuessInput = ({ addGuess, autocomplete, myWord }) => {
   )
 }
 
-GuessInput.propTypes = {
+const WARNINGS = {
+  TOO_SHORT: 'Too short',
+  TOO_LONG: 'Too long',
+  NO_DUPLICATES: 'No duplicates',
+  LETTERS_SIZE: 'Must be between 0-4',
+}
+
+UserInput.propTypes = {
   addGuess: PropTypes.func.isRequired,
   autocomplete: PropTypes.bool,
   myWord: PropTypes.string,
 }
 
-export { GuessInput }
+export { UserInput }
