+++ src/components/input/user-input.test.js
@@ -3,35 +3,26 @@ import React from 'react'
 import { render } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 
-import { GuessInput } from './guess-input'
+import { UserInput } from './user-input'
 
 test('shows placeholder text and disabled submit button when no word is entered', () => {
   const { getByPlaceholderText, getByTestId } = render(
-    <GuessInput addGuess={jest.fn()} />,
+    <UserInput addGuess={jest.fn()} />,
   )
 
   expect(getByPlaceholderText('Enter word...')).toBeInTheDocument()
   expect(getByTestId('letters_input')).toBeInTheDocument()
-  expect(getByTestId('letters_input')).toBeEnabled()
   expect(getByTestId('submit_button')).toBeInTheDocument()
-  expect(getByTestId('submit_button')).toBeDisabled()
-})
-
-test('disables button when word is invalid', () => {
-  const { getByPlaceholderText, getByTestId } = render(
-    <GuessInput addGuess={jest.fn()} />,
-  )
-
-  userEvent.type(getByPlaceholderText('Enter word...'), 'wood')
   expect(getByTestId('submit_button')).toBeDisabled()
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'word')
+
   expect(getByTestId('submit_button')).toBeEnabled()
 })
 
 test('shows warning when word not long enough', () => {
   const { getByPlaceholderText, getByText, queryByText } = render(
-    <GuessInput addGuess={jest.fn()} />,
+    <UserInput addGuess={jest.fn()} />,
   )
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'wor')
@@ -45,7 +36,7 @@ test('shows warning when word not long enough', () => {
 
 test('shows warning when word too long', () => {
   const { getByPlaceholderText, getByText, queryByText } = render(
-    <GuessInput addGuess={jest.fn()} />,
+    <UserInput addGuess={jest.fn()} />,
   )
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'words')
@@ -59,7 +50,7 @@ test('shows warning when word too long', () => {
 
 test('shows warning when duplicates present', () => {
   const { getByPlaceholderText, getByText, queryByText } = render(
-    <GuessInput addGuess={jest.fn()} />,
+    <UserInput addGuess={jest.fn()} />,
   )
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'wood')
@@ -71,9 +62,23 @@ test('shows warning when duplicates present', () => {
   expect(queryByText('No duplicates')).toBeNull()
 })
 
+test('disables button when warning is present', () => {
+  const { getByPlaceholderText, getByText, getByTestId } = render(
+    <UserInput addGuess={jest.fn()} />,
+  )
+
+  userEvent.type(getByPlaceholderText('Enter word...'), 'wood')
+
+  expect(getByText('No duplicates')).toBeInTheDocument()
+  expect(getByTestId('submit_button')).toBeDisabled()
+
+  userEvent.type(getByPlaceholderText('Enter word...'), 'word')
+  expect(getByTestId('submit_button')).toBeEnabled()
+})
+
 test('user can only enter numbers between 0 and 4', () => {
   const { getByPlaceholderText, getByText, getByTestId, queryByText } = render(
-    <GuessInput addGuess={jest.fn()} />,
+    <UserInput addGuess={jest.fn()} />,
   )
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'word')
@@ -111,7 +116,7 @@ test('calls add guess with correct guess format on submit', () => {
   const addGuess = jest.fn()
 
   const { getByPlaceholderText, getByTestId } = render(
-    <GuessInput addGuess={addGuess} />,
+    <UserInput addGuess={addGuess} />,
   )
 
   userEvent.type(getByPlaceholderText('Enter word...'), 'word')
@@ -129,7 +134,7 @@ test('calls add guess with correct guess format on submit', () => {
 
 test('letters is auto-populated and not editable if autocomplete is set to true', () => {
   const { getByPlaceholderText, getByTestId } = render(
-    <GuessInput addGuess={jest.fn()} autocomplete myWord='word' />,
+    <UserInput addGuess={jest.fn()} autocomplete myWord='word' />,
   )
 
   expect(getByTestId('letters_input')).toBeDisabled()
